requirements.txt
python-telegram-bot==20.7
pandas
numpy
yfinance
ta
requests
lightgbm
joblib
